requirements.txt
appnope==0.1.0
argcomplete==1.6.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.5.1
chardet==2.3.0
cycler==0.10.0
decorator==4.0.10
docx2txt==0.6
EbookLib==0.15
enum34==1.1.6
futures==3.0.5
gi==1.2
ipython==5.1.0
ipython-genutils==0.1.0
livestreamer==1.12.2
lxml==3.6.4
matplotlib==1.5.3
nose==1.3.7
numpy==1.11.2
pathlib2==2.1.0
pbr==1.9.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.4.2
prompt-toolkit==1.0.9
ptyprocess==0.5.1
Pygments==2.1.3
pyparsing==2.1.10
pytesseract==0.1.6
python-dateutil==2.6.0
python-pptx==0.6.1
pytz==2016.7
requests==2.11.1
scipy==0.18.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
SpeechRecognition==3.4.6
stevedore==1.12.0
textract==1.5.0
traitlets==4.3.1
virtualenv==15.0.1
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
wcwidth==0.1.7
xlrd==1.0.0
XlsxWriter==0.9.3
